Moved ingest_data() to download_data.py
