+++ src/datapipeline.py
@@ -2,25 +2,8 @@
 Functions to ingest and process data
 """
 import zipfile
-import requests
 
-def ingest_data():
-    """
-    Function to download file from URL
-    """
-    file_url = "https://archive.ics.uci.edu/static/public/352/online+retail.zip"
-
-    # Send an HTTP GET request to the URL
-    response = requests.get(file_url, timeout=30)
-
-    # Check if the request was successful (status code 200)
-    if response.status_code == 200:
-        # Save file to data
-        with open("data/data.zip", "wb") as file:
-            file.write(response.content)
-        print("File downloaded successfully.")
-    else:
-        print(f"Failed to download the file. Status code: {response.status_code}")
+from .download_data import ingest_data
 
 
 def unzip_file():
@@ -38,5 +21,5 @@ def unzip_file():
 
 
 if __name__ == "__main__":
-    ingest_data()
+    ZIPFILE_PATH = ingest_data("https://archive.ics.uci.edu/static/public/352/online+retail.zip")
     unzip_file()
